add middleware folder structure
