+++ middleware/index.js
@@ -1,0 +1,4 @@
+module.exports = {
+    userMiddleware: require('./user'),
+    expensesCategoryMiddleware: require('./expensesCategory')
+}
