+++ routes/user.js
@@ -1,5 +1,6 @@
 const express = require('express');
 const userController = require('@/controllers').user
+const { userMiddleware } = require('@/middleware')
 
 const router = express.Router();
 
@@ -7,10 +8,10 @@ router.post('/register', userController.createUser)
 router.post('/login', userController.verifyUser)
 router
   .route('/:id')
-  .get(userController.validateUser, userController.getUser)
-  .put(userController.validateUser, userController.updateUser)
+  .get(userMiddleware.validateUser, userController.getUser)
+  .put(userMiddleware.validateUser, userController.updateUser)
 
-router.put('/:id/password', userController.validateUser, userController.updatePassword)
+router.put('/:id/password', userMiddleware.validateUser, userController.updatePassword)
 //   .delete(userController.deleteUser);
 
 module.exports = router;
